+++ src/main/java/com/degressly/proxy/mysql/parser/impl/text/IntLenEnc.java
@@ -13,7 +13,7 @@ public class IntLenEnc implements FieldDecoder, FieldEncoder {
 
 	@Override
 	public Pair<Object, Integer> decode(MySQLPacket packet, int offset) {
-		Pair<Integer, Integer> intLengthSizePair = Utils.calculateIntLenEnc(packet.getBody(), offset);
+		Pair<Long, Integer> intLengthSizePair = Utils.calculateIntLenEnc(packet.getBody(), offset);
 		return Pair.of(intLengthSizePair.getLeft(), offset + intLengthSizePair.getRight());
 	}
